TCIC2-6 Made the shoutout command basically as good as it used to be
