+++ _scripts/rustfmt.sh
@@ -3,4 +3,5 @@
 rustfmt --edition 2021 */src/*.rs
 rustfmt --edition 2021 */src/**/*.rs
 rustfmt --edition 2021 */src/**/**/*.rs
+rustfmt --edition 2021 */src/**/**/**/*.rs
 # add more /**/ as necessary
