+++ api_consumers/src/twitch/mod.rs
@@ -1,4 +1,5 @@
 //! Twitch API calls
 
 pub mod auth; // Twitch Auth Routes
-pub mod users; // Twitch User-Realted Lookups
+pub mod channel;
+pub mod users; // Twitch User Related Lookups // Twitch Channel Related Lookups
