+++ api_consumers/src/twitch/users.rs
@@ -1,15 +1,11 @@
 //! Calls to https://api.twitch.tv/helix/users
 
 use reqwest::{Client, Response};
-use std::env;
 
-use database::twitch_access_token;
-use utils;
+use utils::{self, twitch::client_and_access_token};
 
 pub async fn lookup_user_from_login(login: &str) -> Response {
-    let client_id =
-        env::var("TWITCH_CLIENT_ID").expect("Missing TWITCH_CLIENT_ID environment variable.");
-    let access_token = twitch_access_token().unwrap();
+    let (client_id, access_token) = client_and_access_token();
     Client::new()
         .get(format!("https://api.twitch.tv/helix/users?login={}", login))
         .header("Authorization", format!("Bearer {}", access_token))
