+++ utils/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+
+database = { path="../database" }
 rand = "0.8.5"
 twitch-irc = "5.0.1"
 url = "2.5.0"
